Støtter header-parameter for å wipe hele cachen
