+++ server/server.js
@@ -2,7 +2,7 @@ require('dotenv').config();
 
 const express = require('express');
 const next = require('next');
-const { invalidateCachedPage } = require('./invalidate-cached-page');
+const { invalidateCachedPage, wipePageCache } = require('./incremental-cache');
 const { initHeartbeat } = require('./revalidator-proxy-heartbeat.js');
 
 const app = next({ dev: process.env.NODE_ENV !== 'production' });
@@ -20,10 +20,16 @@ app.prepare().then(() => {
     }
 
     server.all('*', (req, res) => {
-        const { secret, invalidate } = req.headers;
+        const { secret, invalidate, wipeCache } = req.headers;
+
         if (invalidate && secret === serviceSecret) {
             invalidateCachedPage(req.path, app);
             return res.status(200).send(`Invalidating cache for ${req.path}`);
+        }
+
+        if (wipeCache && secret === serviceSecret) {
+            wipePageCache(app);
+            return res.status(200).send('Wiping page cache');
         }
 
         return handle(req, res);
